move out serializer tests
